+++ src/main/webapp/resources/js/pages/UserGroupsPage/index.js
@@ -1,23 +1,43 @@
-import React from "react";
+import React, { lazy, Suspense } from "react";
 import { render } from "react-dom";
-import { PageWrapper } from "../../components/page/PageWrapper";
-import { PagedTableProvider } from "../../components/ant.design/PagedTable";
-import { UserGroupsTable } from "./components/UserGroupsTable";
-import { setBaseUrl } from "../../utilities/url-utilities";
+import { Router } from "@reach/router";
+import { Spin } from "antd";
+
+const UserGroupsPage = lazy(() => import("./components/UserGroupsPage"));
+const UserGroupsDetailsPage = lazy(() =>
+  import("./components/UserGroupDetailsPage")
+);
 
 /**
- * React component to display pages related to User Groupss
+ * React component to display pages related to User Groups.  This is a base page
+ * for both listing of user groups and user group details.
  * @returns {*}
  * @constructor
  */
-export function Index() {
+export function UserGroups() {
   return (
-    <PagedTableProvider url={setBaseUrl(`/ajax/user-groups/list`)}>
-      <PageWrapper title={i18n("UserGroupsPage.title")}>
-        <UserGroupsTable />
-      </PageWrapper>
-    </PagedTableProvider>
+    <Suspense
+      fallback={
+        <div
+          style={{
+            height: "100%",
+            display: "flex",
+            justifyContent: "center",
+            alignItems: "center",
+          }}
+        >
+          <span>
+            <Spin /> Fetching important data
+          </span>
+        </div>
+      }
+    >
+      <Router style={{ height: "100%" }}>
+        <UserGroupsPage path="/groups" />
+        <UserGroupsDetailsPage path="/groups/:id" />
+      </Router>
+    </Suspense>
   );
 }
 
-render(<Index />, document.querySelector("#groups-root"));
+render(<UserGroups />, document.querySelector("#groups-root"));